Add files via upload
We created a program using JAVA for a coffee machine. The device offers:
short expresso
extended expresso
cappuccino
The appliance also has a water container, a coffee container, and a milk container.
When preparing each drink, the device uses water, coffee and sometimes milk, depending on the drink.
If there is no more water, the device should display the message: "Out of watter".
If there is no coffee: "Machine is out of coffee".
When there is no milk: "Machine is out of milk".
During each preparation of the drink, the device must display a message about what it is currently doing, e.g. pourWater, pourMilk.
For each ordered drink, the device deducts the value from the total inserted credit, if there is no credit, the device must output the message: "Not enough credit".